Move api.js to utils folder
