+++ components/Map.js
@@ -1,7 +1,9 @@
 import React from 'react';
-import { MapView, Location, Permissions, Marker, TouchableHighlight, View, Text } from 'expo';
+import {
+  MapView, Location, Permissions, Marker, TouchableHighlight, View, Text,
+} from 'expo';
 
-import * as api from '../api';
+import * as api from '../utils/api';
 
 export default class Map extends React.Component {
   state = {
@@ -16,7 +18,7 @@ export default class Map extends React.Component {
     markers: [],
     pins: [],
   };
-  
+
   render() {
     return (
       <MapView
@@ -27,7 +29,7 @@ export default class Map extends React.Component {
           latitudeDelta: 0.0922,
           longitudeDelta: 0.0421,
         }}
-        showsUserLocation={true}
+        showsUserLocation
       >
         {this.state.pins.map(pin => (
           <MapView.Marker
@@ -35,10 +37,9 @@ export default class Map extends React.Component {
             coordinate={{ latitude: Number(pin.latitude), longitude: Number(pin.longitude) }}
             title={pin.creator}
             description={pin.timestamp}
-            onPress={() =>
-              this.props.navigation.navigate('Pin', {
-                pin: pin,
-              })
+            onPress={() => this.props.navigation.navigate('Pin', {
+              pin,
+            })
             }
           />
         ))}
@@ -52,7 +53,7 @@ export default class Map extends React.Component {
   }
 
   _getLocationAsync = async () => {
-    let { status } = await Permissions.askAsync(Permissions.LOCATION);
+    const { status } = await Permissions.askAsync(Permissions.LOCATION);
     if (status !== 'granted') {
       this.setState({
         locationResult: 'Permission to access location was denied',
